script: Die class
"Holy Chits" is a joke based on the mechanism of "chits" that were
used for random number generation in table-top games back before
polyhedral dice (that weren't the traditional cube) were commonplace.
They worked as basic as you'd expect: you'd photo-copy a page out your
gamebook, snip out the small grid, sort them into little jars or bowls
based on the number range on the back, and draw a number.

Nowadays people *really* like the clatter of their dice on the table.
Especially if they're the ones from your Pathfinder group that had a
tendency of slinging spells around that called for at least ten dice.
But Holy Chits! is a project of discretion; it isn't necessarily a
throwback to the use of chits, but instead it serves as a pre-rolled
list of numbers to print out and tuck in your binder in case you lent
your dice to your friends who don't have any, forgot to bring your
set, or are not in a situation where pulling out dice and rolling them
would be acceptable.