feature: status text changes on hovering elements
Fun fact: custom data is a thing.  All in all, this sounds like it'd
be a good way to "teach" a user how to use it by that.  Then again
it's way there out of sight and hopefully it would be more obvious...
But let's say this is groundwork being laid for whatever.

https://www.w3.org/TR/html51/dom.html#custom-data-attribute